+++ lib/src/cli/formatter_options.dart
@@ -39,6 +39,11 @@ class FormatterOptions {
   /// Sets the exit code to 1 if any changes are made.
   final bool setExitIfChanged;
 
+  /// Flags to enable experimental language features.
+  ///
+  /// See dart.dev/go/experiments for details.
+  final List<String> experimentFlags;
+
   FormatterOptions(
       {this.indent = 0,
       this.pageWidth = 80,
@@ -47,8 +52,10 @@ class FormatterOptions {
       this.show = Show.changed,
       this.output = Output.write,
       this.summary = Summary.none,
-      this.setExitIfChanged = false})
-      : fixes = [...?fixes];
+      this.setExitIfChanged = false,
+      List<String>? experimentFlags})
+      : fixes = [...?fixes],
+        experimentFlags = [...?experimentFlags];
 
   /// Called when [file] is about to be formatted.
   ///