+++ lib/src/io.dart
@@ -29,7 +29,8 @@ Future<void> formatStdin(
     var formatter = DartFormatter(
         indent: options.indent,
         pageWidth: options.pageWidth,
-        fixes: options.fixes);
+        fixes: options.fixes,
+        experimentFlags: options.experimentFlags);
     try {
       options.beforeFile(null, name);
       var source = SourceCode(input.toString(),
@@ -138,7 +139,8 @@ bool processFile(FormatterOptions options, File file, {String? displayPath}) {
   var formatter = DartFormatter(
       indent: options.indent,
       pageWidth: options.pageWidth,
-      fixes: options.fixes);
+      fixes: options.fixes,
+      experimentFlags: options.experimentFlags);
   try {
     var source = SourceCode(file.readAsStringSync(), uri: file.path);
     options.beforeFile(file, displayPath);